fixed some compiler warnings
git-svn-id: svn://orca.physics.unc.edu/Orca@7695 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/GPIB/Pulser33500/ORPulser33500Chan.h
@@ -133,7 +133,7 @@ typedef struct Pulser33500CustomWaveformStruct {
 - (void) writeFrequency;
 - (void) writeBurstRate:(float)aValue;
 - (void) writeBurstRate;
-- (void) writeBurstCount:(float)aValue;
+- (void) writeBurstCount:(int)aValue;
 - (void) writeBurstCount;
 - (void) writeBurstPhase;
 - (void) writeTriggerSource;
